[TS-379] Refactor: habit record name refactoring
