+++ src/constants/path.ts
@@ -11,7 +11,7 @@ export const PATH = {
 		`/rest-record/${habitId}/${year}/${month}/${date}`,
 	PRACTICE_RECORD: (habitId: string, year: string, month: string, date: string) =>
 		`/practice-record/${habitId}/${year}/${month}/${date}`,
-	HABIT_RECORD: (habitId: number, year: number, month: number, date: number) =>
+	HABIT_RECORD: (habitId: string, year: string, month: string, date: string) =>
 		`/habit-record/${habitId}/${year}/${month}/${date}`,
 	EDIT_HABIT: (habitId?: string) => `/habit-edit/${habitId}`,
 	DELETE_HABIT: (habitId?: string) => `/habit-delete/${habitId}`,
